virtualClassroom/requirement.txt
Package                       Version
----------------------------- -------
asgiref                       3.4.1  
Django                        3.2.7  
djangorestframework           3.12.4 
djangorestframework-simplejwt 4.8.0  
pip                           20.0.2 
pkg-resources                 0.0.0  
PyJWT                         2.1.0  
pytz                          2021.1 
setuptools                    44.0.0 
sqlparse                      0.4.1  
wheel                         0.34.2 
